Feat: 번호 인증 버튼 삭제 및 '위와 동일하게 채우기' 기능 구현
