+++ src/pages/Order/components/index.js
@@ -1,6 +1,5 @@
 import OrderPersonContainer from "./OrderPersonContainer";
 import OrderStickyContainer from "./OrderStickyContainer";
-import DestinationContainer from "./DestinationContainer";
 import OrderProduct from "./OrderProduct";
 import CouponContainer from "./CouponContainer";
 import PointContainer from "./PointContainer";
@@ -9,7 +8,6 @@ import SelectedPayOption from "./SelectedPayOption";
 export {
     OrderPersonContainer,
     OrderStickyContainer,
-    DestinationContainer,
     OrderProduct,
     CouponContainer,
     PointContainer,
